requirements.txt
asttokens==2.4.1
attrs==23.2.0
decorator==5.1.1
executing==2.0.1
fancycompleter==0.9.1
ipython==8.26.0
jedi==0.19.1
matplotlib-inline==0.1.7
parso==0.8.4
pdbpp==0.10.3
pexpect==4.9.0
pillow==10.3.0
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
pyrepl==0.9.0
six==1.16.0
stack-data==0.6.3
traitlets==5.14.3
wcwidth==0.2.13
wmctrl==0.5
